# More optimizations and corrections