--- a/src/Elm/Parser/Expression.elm
+++ b/src/Elm/Parser/Expression.elm
@@ -971,7 +971,7 @@ tupledExpressionIfNecessaryFollowedByRecordAccess =
 
 allowedPrefixOperatorFollowedByClosingParensOneOf : Parser (WithComments (Node Expression))
 allowedPrefixOperatorFollowedByClosingParensOneOf =
-    ParserFast.whileWithoutLinebreakAnd2PartUtf16ValidateMapWithRangeFollowedBySymbol
+    ParserFast.whileWithoutLinebreakAnd2PartUtf16ValidateMapWithRangeBacktrackableFollowedBySymbol
         (\operatorRange operator ->
             { comments = Rope.empty
             , syntax =