… CHANGELOG.md
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/ParserFast.elm
… src/ParserWithComments.elm
… tests/Elm/Parser/CaseExpressionTests.elm
… tests/Elm/Parser/CommentTest.elm
… tests/Elm/Parser/ExposeTests.elm
… tests/Elm/Parser/ExpressionTests.elm
… tests/Elm/Parser/LayoutTests.elm
… tests/Elm/Parser/ModuleTests.elm
… tests/Elm/Parser/NumbersTests.elm
… tests/Elm/Parser/ParserWithCommentsTestUtil.elm
… tests/Elm/Parser/PatternTests.elm
… tests/Elm/Parser/TestUtil.elm
… tests/Elm/Parser/TokenTests.elm
… tests/Elm/Parser/TypeAnnotationTests.elm
… tests/Elm/ProcessingTests.elm
… tests/Elm/WriterTests.elm