Contains a bunch of optimizations and fixes, including for #258 which has arguably been the biggest improvement (the commit https://github.com/stil4m/elm-syntax/pull/260/commits/33fe359970f6df61b9e187854309a7d8a34a4314)

- before: 2850.800000000745 ms to do 100 runs => 28.50800000000745 ms/run (likely a bit slower than after of last PR due to regression fixes and better averaging)
- after: 1933 ms to do 100 runs => 19.33 ms/run